requirements.txt
altgraph==0.17.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
packaging==25.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pyparsing==3.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==78.1.1
uritemplate==4.1.1
urllib3==2.4.0
